+++ qt/src/main/java/org/bytedeco/qt/presets/Qt5Core.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 Greg Hart, Samuel Audet
+ * Copyright (C) 2019-2020 Greg Hart, Samuel Audet
  *
  * Licensed either under the Apache License, Version 2.0, or (at your option)
  * under the terms of the GNU General Public License as published by
@@ -27,10 +27,12 @@ import java.util.Map;
 import org.bytedeco.javacpp.Loader;
 import org.bytedeco.javacpp.annotation.Platform;
 import org.bytedeco.javacpp.annotation.Properties;
+import org.bytedeco.javacpp.presets.javacpp;
 import org.bytedeco.javacpp.tools.Info;
 import org.bytedeco.javacpp.tools.InfoMap;
 
 @Properties(
+    inherit = javacpp.class,
     target = "org.bytedeco.qt.Qt5Core",
     global = "org.bytedeco.qt.global.Qt5Core",
     helper = "org.bytedeco.qt.helper.Qt5Core",