+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Add dependency on presets for `jnijavacpp` and `javacpp-platform` artifact to fix issues at load time ([issue bytedeco/javacv#1305](https://github.com/bytedeco/javacv/issues/1305))
  * Bundle the official Java API of ONNX Runtime via the `jnionnxruntime` library
  * Add CUDA-enabled build for ONNX Runtime via `-gpu` extension
  * Fix presets for LLVM 9.0 where libclang would fail to load on Windows ([issue #830](https://github.com/bytedeco/javacpp-presets/issues/830))